requirements/base.txt
Django==4.2
djangorestframework==3.15.2
drf-standardized-errors==0.14.1
django-filter==25.1
djangorestframework_simplejwt==5.5.0
django-jazzmin==3.0.1

drf-spectacular==0.28.0

django-cors-headers==4.7.0

django-extensions==3.1.3

django-environ==0.12.0

django-cors-headers==4.7.0

gunicorn==23.0.0

django-polymorphic==3.1.0
setuptools==80.3.1

# firebase
firebase-admin==6.7.0

# image processing
Pillow==10.1.0

# email
sendgrid==6.11.0


openpyxl==3.1.5

# sentry
sentry-sdk==2.27.0
